execute calls to teams in already running container
check if a teams container is already running and exec a call to teams
in it -> this allows to open deeplinks with the running container

+++ teams.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# check if teams container is running and use existing container if so
+if docker exec teams true > /dev/null 2>&1; then
+	docker exec teams teams "$@"
+	exit $?
+fi
+
 # ensure xdg socket is setup correctly
 rm -f ~/x11docker/teams/.xdg.sock
 mkfifo ~/x11docker/teams/.xdg.sock
